backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
motor==3.6.0
numpy
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.30.6
pyjwt
